requirements.txt
absl-py==2.1.0
albumentations==1.4.3
annotated-types==0.6.0
anyio==4.9.0
astroid==3.1.0
astunparse==1.6.3
Babel==2.14.0
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
crashtest==0.4.1
cycler==0.12.1
dill==0.3.8
distlib==0.3.9
dulwich==0.22.8
exceptiongroup==1.2.0
-e git+https://github.com/ankandrew/fast-plate-ocr.git@49954297c2226ad6c488f6fbfb93169b579b815d#egg=fast_plate_ocr
fastjsonschema==2.21.1
filelock==3.13.4
findpython==0.6.3
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
ghp-import==2.1.0
google-pasta==0.2.0
griffe==0.42.1
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.3
joblib==1.4.0
keras==3.2.1
keyring==25.6.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mike==2.0.0
mkdocs==1.5.3
mkdocs-autorefs==1.0.1
mkdocs-material==9.5.17
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.3
mkdocstrings-python==1.9.2
ml-dtypes==0.3.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
mypy==1.10.0
mypy-extensions==1.0.0
namex==0.0.7
networkx==3.3
numpy==1.26.4
onnx==1.16.0
onnxruntime==1.19.2
onnxsim==0.4.36
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
paginate==0.5.6
pandas==2.2.2
pandas-stubs==2.2.1.240316
pathspec==0.12.1
pbs-installer==2025.3.11
pillow==10.3.0
pkginfo==1.12.1.2
platformdirs==4.2.0
pluggy==1.5.0
poetry==2.1.1
poetry-core==2.1.1
protobuf==3.20.3
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pylint==3.1.0
pymdown-extensions==10.7.1
pyparsing==3.1.2
pyproject_hooks==1.2.0
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
RapidFuzz==3.12.2
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
ruff==0.5.2
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf2onnx==1.16.1
threadpoolctl==3.4.0
tifffile==2024.2.12
tomli==2.0.1
tomlkit==0.12.4
torch==2.2.2
tqdm==4.66.2
trove-classifiers==2025.3.13.13
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
verspec==0.1.0
virtualenv==20.29.3
watchdog==4.0.0
Werkzeug==3.0.2
wrapt==1.16.0
xattr==1.1.4
zipp==3.18.1
zstandard==0.23.0
